+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Automatic check for missing dependencies of assembly addons on *assembly sync*.
 - Ability to use existing assembly for project via `odood assembly use` command.
   This could be useful in CI to automate assembly sync process.
+- Added new options to specify commit params for `odood assembly sync` command.
 
 ### Changed
 - `odood log` now will automatically show the end of logfile